+++ src/interfaces/IOracle.sol
@@ -18,6 +18,6 @@
 pragma solidity 0.8.17;
 
 interface IOracle {
-    function fee(uint32 toChainId, address ua) external view returns (uint256);
-    function assign(uint256 index, uint32 toChainId, address ua) external payable returns (uint256);
+    function fee(uint256 toChainId, address ua) external view returns (uint256);
+    function assign(uint256 index, uint256 toChainId, address ua) external payable returns (uint256);
 }
