+++ src/interfaces/IRelayer.sol
@@ -18,8 +18,8 @@
 pragma solidity 0.8.17;
 
 interface IRelayer {
-    function fee(uint32 toChainId, address ua, uint256 size, bytes calldata params) external view returns (uint256);
-    function assign(uint256 index, uint32 toChainId, address ua, uint256 size, bytes calldata params)
+    function fee(uint256 toChainId, address ua, uint256 size, bytes calldata params) external view returns (uint256);
+    function assign(uint256 index, uint256 toChainId, address ua, uint256 size, bytes calldata params)
         external
         payable
         returns (uint256);
